requirements.txt
Faker
python-dotenv
ipykernel
aiokafka